refactor(studio): consolidate GitHub types into single types.ts file
Merge shared-types.ts and content-metadata-schema.ts into types.ts for better organization
