+++ apps/studio.giselles.ai/lib/vector-stores/github/index.ts
@@ -2,4 +2,4 @@ export { getGitHubVectorStores } from "./get-github-vector-stores";
 export { fetchIngestTargets } from "./ingest/fetch-ingest-targets";
 export { processRepository } from "./ingest/process-repository";
 export { gitHubQueryService } from "./query/service";
-export type { RepositoryWithStatuses } from "./shared-types";
+export type { RepositoryWithStatuses } from "./types";
