--- apps/studio.giselles.ai/lib/vector-stores/github/shared-types.ts
@@ -1,12 +1,0 @@
-import type {
-	githubRepositoryContentStatus,
-	githubRepositoryIndex,
-} from "@/drizzle";
-
-/**
- * Repository with all its content statuses
- */
-export type RepositoryWithStatuses = {
-	repository: typeof githubRepositoryIndex.$inferSelect;
-	contentStatuses: (typeof githubRepositoryContentStatus.$inferSelect)[];
-};
